Added logout and more comments

+++ goods/views.py
@@ -3,13 +3,16 @@ from django.http import HttpResponseNotAllowed, HttpResponseRedirect, HttpRespon
 from .models import Good, Category
 from .forms import CategoryForm, GoodForm, SearchForm, RegistrationForm, LoginForm
 from django.urls import reverse
-from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth import authenticate, login
 from django.contrib.auth.decorators import login_required
 
 
+#checking if current user is creator 
 def can_modify_object(user, obj):
     return user.is_authenticated and obj.creator == user
 
+
+#view of regitration
 def register_user(request):
     if request.method == 'POST':
         form = RegistrationForm(request.POST)
@@ -25,6 +28,8 @@ def register_user(request):
         form = RegistrationForm()
     return render(request, 'authentication/register.html', {'form': form})
 
+
+#view of login for users
 def login_user(request):
     error_message = None  # Initialize error message as None
     
@@ -45,9 +50,6 @@ def login_user(request):
     return render(request, 'authentication/login.html', {'form': form, 'error_message': error_message})
 
 
-def logout_user(request):
-    return logout(request)
-
 @login_required(login_url='login_user')
 #view of home page with list of all goods and search of goods by name
 def get_goods(request):
@@ -66,6 +68,7 @@ def get_goods(request):
 
     return render(request, template_name='goods/goods.html', context={'goods': goods, 'form': form})
 
+
 @login_required(login_url='login_user')
 #view of one good item with mode details and full picture
 def get_good(request, pk=None):
@@ -74,12 +77,12 @@ def get_good(request, pk=None):
         return render(request, template_name='goods/good.html', context={'good': good})
     return HttpResponseNotAllowed(["GET"])
 
+
 @login_required(login_url='login_user')
 #view of editing good item
 def update_good(request, pk=None):
     obj = get_object_or_404(Good, pk=pk)
     good = Good.objects.get(pk=pk)
-
     # Check if the current user has permission to modify the object
     if not can_modify_object(request.user, obj):
         return HttpResponseForbidden("You don't have permission to modify this object.")
@@ -92,13 +95,19 @@ def update_good(request, pk=None):
         form = GoodForm(instance=good)
     return render(request, 'goods/update_good.html', {'form': form})
 
+
 @login_required(login_url='login_user')
 #view of delete good item 
 def delete_good(request, pk=None):
+    obj = get_object_or_404(Good, pk=pk)
+    # Check if the current user has permission to modify the object
+    if not can_modify_object(request.user, obj):
+        return HttpResponseForbidden("You don't have permission to modify this object.")
     if request.method == 'GET':
         Good.objects.get(pk=pk).delete()
         return HttpResponseRedirect(reverse('get_goods'))
     return HttpResponseNotAllowed(["GET"])
+
 
 @login_required(login_url='login_user')
 #view of creating new good item 
@@ -113,6 +122,7 @@ def create_good(request):
     else:
         form = GoodForm()
     return render(request, 'goods/create_good.html', {'form': form})
+
 
 @login_required(login_url='login_user')
 #view of page with list of all categories and search of categories by name
@@ -133,6 +143,7 @@ def get_categories(request):
     else:
         return HttpResponseNotAllowed(["GET", "POST"])
 
+
 @login_required(login_url='login_user')
 #view of one category 
 def get_category(request, pk=None):
@@ -141,10 +152,15 @@ def get_category(request, pk=None):
         return render(request, template_name='categories/category.html', context={'category': category}) 
     return HttpResponseNotAllowed(["GET"])
 
+
 @login_required(login_url='login_user')
 #view of edit for category
 def update_category(request, pk=None):
     category = Category.objects.get(pk=pk)
+    obj = get_object_or_404(Category, pk=pk)
+    # Check if the current user has permission to modify the object
+    if not can_modify_object(request.user, obj):
+        return HttpResponseForbidden("You don't have permission to modify this object.")
     if request.method == 'POST':
         form = CategoryForm(request.POST, instance=category)
         if form.is_valid():
@@ -154,13 +170,19 @@ def update_category(request, pk=None):
         form = CategoryForm(instance=category)  
     return render(request, 'categories/update_category.html', {'form': form})
 
+
 @login_required(login_url='login_user')
 #view of dekete category
 def delete_category(request, pk=None):
+    obj = get_object_or_404(Category, pk=pk)
+    # Check if the current user has permission to modify the object
+    if not can_modify_object(request.user, obj):
+        return HttpResponseForbidden("You don't have permission to modify this object.")
     if request.method == 'GET':
         Category.objects.get(pk=pk).delete()
         return HttpResponseRedirect(reverse('get_categories'))
     return HttpResponseNotAllowed(["GET"])
+
 
 @login_required(login_url='login_user')
 #view of craeting new one category
@@ -175,4 +197,3 @@ def create_category(request):
     else:
         form = CategoryForm()
     return render(request, template_name='categories/create_category.html', context={'form': form})
-
